Refactoring, adding documentation.
