+++ src/main/java/com/uber/cadence/largeblob/Future.java
@@ -13,10 +13,14 @@
  *  permissions and limitations under the License.
  */
 
-package com.uber.cadence.internal.largeblob;
+package com.uber.cadence.largeblob;
 
 import java.io.IOException;
 
+/**
+ * Future is used for passing around potentially large parameters between activities. Future can never be used in Workflow code because it uses an external storage and that will make
+ * workflow code non deterministic. Small amounts of data are stored in the future instance itself, while larger amounts of data are stored in an external storage.
+ */
 public class Future {
 
   private byte[] encoded;
@@ -24,10 +28,23 @@ public class Future {
   private final Storage storage;
   private final long maxBytesInMemory;
 
+  public Future(Storage storage, long maxBytesInMemory, byte[] encoded) {
+    this.storage = storage;
+    this.maxBytesInMemory = maxBytesInMemory;
+    this.encoded = encoded;
+  }
+
+  public Future(Storage storage, long maxBytesInMemory, String url) {
+    this.url = url;
+    this.storage = storage;
+    this.maxBytesInMemory = maxBytesInMemory;
+  }
+
   public Future(Storage storage, long maxBytesInMemory) {
     if (storage == null) {
       throw new IllegalArgumentException("storage can't be null");
     }
+
     this.storage = storage;
     this.maxBytesInMemory = maxBytesInMemory;
   }
@@ -59,5 +76,13 @@ public class Future {
     } else {
       this.encoded = null;
     }
+  }
+
+  public byte[] getEncoded() {
+    return encoded;
+  }
+
+  public String getUrl() {
+    return url;
   }
 }
