+++ src/main/java/com/uber/cadence/largeblob/Storage.java
@@ -13,15 +13,35 @@
  *  permissions and limitations under the License.
  */
 
-package com.uber.cadence.internal.largeblob;
+package com.uber.cadence.largeblob;
 
 import java.io.IOException;
 
+/**
+ * Storage is an abstraction for storing large parameters to access inside of activities.0
+ */
 public interface Storage {
 
+  /**
+   * Gets the data based on uri provided
+   * @param uri uri.
+   * @return the data as a byte array.
+   * @throws IOException should be thrown in any implementation class in case of problems accessing the datastore.
+   */
   byte[] get(String uri) throws IOException;
 
+  /**
+   * Stores data based on uri provided.
+   * @param uri uri.
+   * @param bytes bytes.
+   * @throws IOException should be thrown in any implementation class in case of problems with the datastore
+   */
   void put(String uri, byte[] bytes) throws IOException;
 
+  /**
+   * Deletes data based on uri provided.
+   * @param uri uri.
+   * @throws IOException should be thrown in any implementation class in case of problems with the datastore
+   */
   void delete(String uri) throws IOException;
 }
